requirements.txt
SQLAlchemy
PyMySQL
pandas
XlsxWriter
python-dotenv
openpyxl
dashscope
pytest

fastapi
uvicorn
httpx